Refactor (Player): Group player-related headers and sources into `src/player`
