+++ src/renderer/renderer.cpp
@@ -1,5 +1,5 @@
 #include <renderer/renderer.h>
-#include <player.h>
+#include <player/player.h>
 #include <map.h>
 #include <raycaster.h>
